requirements.txt
qrcode==6.1
Flask==1.1.1
gunicorn==20.0.4
Pillow==8.1.1
